-- relaxing the signals for play/pause/resume for the non RTMP streams -- promoted GetDecodedBytesCount to 64 bit value
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@496 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/include/protocols/baseprotocol.h
@@ -210,7 +210,7 @@ public:
 	/*!
 		@brief Get the total amount of bytes that this protocol transported (raw bytes).
 	*/
-	virtual uint32_t GetDecodedBytesCount();
+	virtual uint64_t GetDecodedBytesCount();
 
 	/*!
 		@brief This function can be called by anyone who wants to signal the transport layer that there is data ready to be sent. pExtraParameters usually is a pointer to an OutboundBuffer, but this depends on the protocol type.
